fix(webhook test): update webhook unit test
