--- zanata-war/src/test/java/org/zanata/service/impl/WebhookPublisherCallback.java
@@ -1,55 +1,0 @@
-/*
- * Copyright 2015, Red Hat, Inc. and individual contributors as indicated by the
- * @author tags. See the copyright.txt file in the distribution for a full
- * listing of individual contributors.
- *
- * This is free software; you can redistribute it and/or modify it under the
- * terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This software is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this software; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA, or see the FSF
- * site: http://www.fsf.org.
- */
-
-package org.zanata.service.impl;
-
-import org.mockserver.mock.action.ExpectationCallback;
-import org.mockserver.model.Header;
-import org.mockserver.model.HttpRequest;
-import org.mockserver.model.HttpResponse;
-
-import static org.mockserver.model.HttpResponse.response;
-
-/**
- * This is to return same header and body from httpRequest to the response
- *
- * See {@link WebhooksPublisherITCase}
- *
- * @author Alex Eng <a href="mailto:aeng@redhat.com">aeng@redhat.com</a>
- */
-public class WebhookPublisherCallback implements
-        ExpectationCallback {
-
-    public static String LISTENER_PATH = "/listener";
-
-    @Override
-    public HttpResponse handle(HttpRequest httpRequest) {
-        if (httpRequest.getPath().getValue().endsWith(LISTENER_PATH)) {
-            HttpResponse response = response().withStatusCode(200)
-                    .withBody(httpRequest.getBody());
-            for (Header header : httpRequest.getHeaders()) {
-                response.withHeader(header);
-            }
-            return response;
-        }
-        return response().withStatusCode(404);
-    }
-}
